fix(flags): unsigned per-bit shift in generated SzFlags.hpp (#13)
tools/gen_flags.py emitted the flag-reflection helpers with
'const int64_t singleBit = (1LL << index)'. At index 63 that shifts a signed
1LL into the sign bit — implementation-defined per [expr.shift] — which cppcheck
reports as shiftTooManyBitsSigned, failing downstream builds that gate on
--enable=portability.

Emit 'static_cast<int64_t>(1ULL << index)' at both sites (FlagsToString and the
group-scoped overload). The shift is now well-defined; the resulting int64 bit
pattern is unchanged on two's-complement targets, so no API or value change.

Reproduce-first verified: cppcheck 2.13 flagged SzFlags.hpp:567 and :599 before,
clean after regeneration. 207/207 tests pass (ASan, real engine).

Closes #13.

+++ tools/gen_flags.py
@@ -439,7 +439,7 @@ inline std::string SzFlags::FlagsToString(SzFlags flag) {{
         appendNames(whole->second);
     }} else {{
         for (int index = 0; index < 64; ++index) {{
-            const int64_t singleBit = (1LL << index);
+            const int64_t singleBit = static_cast<int64_t>(1ULL << index);
             if ((value & singleBit) == 0) {{
                 continue;
             }}
@@ -471,7 +471,7 @@ inline std::string SzFlags::FlagsToString(SzFlagUsageGroup group, SzFlags flag) 
         out += "{{ NONE }}";
     }} else {{
         for (int index = 0; index < 64; ++index) {{
-            const int64_t singleBit = (1LL << index);
+            const int64_t singleBit = static_cast<int64_t>(1ULL << index);
             if ((value & singleBit) == 0) {{
                 continue;
             }}
